implement take2
